requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.13.0
pyyaml>=6.0
tqdm>=4.65.0
rich>=13.0.0
pytest>=7.4.0
